IW-71 Remove Landing and rename Landing 2.0
