+++ src/imports/components/Typography/SectionHeading/index.styl
@@ -4,7 +4,6 @@
   font-family "Roboto Mono", sans-serif
   text-transform uppercase
   text-align center
-  color $white
 
   .sub-title
     opacity 0.4
