+++ src/pages/Landing/index.styl
@@ -1,2 +1,2 @@
-.landing-old
+.landing
   overflow: hidden
