--- src/pages/Landing/sections/Main/components/Description/index.styl
@@ -1,15 +1,0 @@
-@import "~styles/config.styl"
-
-.main-description
-  width 470px
-  min-height 104px
-  margin-bottom 32px
-  color: #4A4A4A
-  line-height 26px
-  opacity 0
-  animation opacityMainSection .5s forwards
-  animation-duration .5s
-  animation-delay 1.6s
-
-  span
-    font-weight $medium-weight
