+++ src/pages/Landing/sections/Main/components/NavigationBar/index.js
@@ -1,25 +1,29 @@
-import React, {Component} from 'react';
-import {Link} from 'react-scroll'
+import React from 'react';
+import {Link} from 'react-scroll';
 
 import './index.styl';
 
-class NavigationBar extends Component {
+const links = [
+  {text: 'Services', to: "services"},
+  {text: 'Process', to: "the-process"},
+  {text: 'Contact', to: "contact-us"},
+];
+
+class NavigationBar extends React.Component {
+  renderLinks() {
+    return links.map(({text, to}, i) => (
+      <Link activeClass="active" className="hover-grow" to={to} smooth duration={500 + (i * 200)}>
+        {text}
+      </Link>
+    ));
+  }
+
   render() {
     return (
       <div className="navigation-bar">
+        <div className="logo" />
         <div className="links">
-          <Link activeClass="active" to="services" smooth duration={500}>
-            Services
-          </Link>
-          <Link activeClass="active" to="the-process" smooth duration={700}>
-            Process
-          </Link>
-          <Link activeClass="active" to="work" smooth duration={900}>
-            Work
-          </Link>
-          <Link activeClass="active" to="contact-us" smooth duration={1100}>
-            Contact
-          </Link>
+          {this.renderLinks()}
         </div>
       </div>
     );
