+++ src/pages/Landing/sections/Main/components/NavigationBar/index.styl
@@ -2,16 +2,26 @@
 
 .navigation-bar
   display flex
-  justify-content flex-end
-  margin-top 32px
-  min-height 20px
+  justify-content space-between
+  align-items center
+  padding-top 40px
+  padding-left 20px
+  padding-right 20px
+
+  +iphone()
+    padding-top: 0
+    padding-left 0
+    display block
+
+  .logo
+    width: 48px
+    height: 49px
+    background: url("../../images/logo@2x.png") no-repeat;
+    background-size: @width @height;
 
   .links
     display flex
-    opacity 0
-    animation opacityMainSection .5s forwards
-    animation-duration .5s
-    animation-delay 1.6s
+    iphoneDisabled()
 
     a
       color $black
@@ -23,10 +33,3 @@
 
       &:last-child
         margin-right 0
-
-@keyframes opacityMainSection
-  0%
-    opacity 0
-
-  100%
-    opacity 1
