+++ src/pages/Landing/sections/Main/images/logo@2x.png
