+++ src/pages/Landing/sections/Main/images/main-art-mobile@2x.png
