+++ src/pages/Landing/sections/Main/index.styl
@@ -1,8 +1,72 @@
-@import "~styles/config.styl"
+@import '~styles/config.styl'
 
-.landing .main-section
-  display flex
-  flex-direction column
-  justify-content center
+.main
+  background url(./images/bg.png)
+  background-size cover
   height 100vh
-  margin 0 124px
+
+  +iphone()
+    padding: 40px
+    background url(./images/main-art-mobile@2x.png) center bottom no-repeat
+    background-size 100% auto
+
+  .main-section
+    display flex
+    flex-direction column
+    max-width 1160px
+    margin 0 auto
+    height: 100%
+    position: relative
+
+    +iphone()
+      max-width: 100%
+      width: 100%
+
+    .main-info-bg
+      background url(./images/main-art.png) right no-repeat
+      background-size 760px 560px
+      margin-top 20px
+      display flex
+      flex 1
+
+      +iphone()
+        background: none
+        margin-top: 0
+        display: block
+
+      .main-info
+        display flex
+        flex-direction column
+        justify-content center
+        min-height 560px
+        width 452px
+        margin-left 20px
+
+        +iphone()
+          width: 100%
+          margin: 0
+          display block
+          margin-top: 32px
+
+        .title
+
+          +iphone()
+            font-size: 21px
+            line-height: 1.5
+            margin-bottom: 16px
+
+        .desc
+          color $dark-gray
+          line-height 32px
+
+          +iphone()
+            line-height: 2
+
+        .buttons
+          display flex
+          margin-top 38px
+
+          iphoneDisabled()
+
+          .explore-btn
+            margin-right 32px
