+++ src/pages/Landing/sections/TheProcess/components/Step/index.js
