+++ src/pages/Landing/sections/Work/components/WorkItem/index.styl
