--- src/pages/Landing2.0/index.styl
@@ -1,2 +1,0 @@
-.landing
-  overflow: hidden
