--- src/pages/Landing2.0/sections/Main/components/NavigationBar/index.styl
@@ -1,35 +1,0 @@
-@import '~styles/config.styl'
-
-.navigation-bar
-  display flex
-  justify-content space-between
-  align-items center
-  padding-top 40px
-  padding-left 20px
-  padding-right 20px
-
-  +iphone()
-    padding-top: 0
-    padding-left 0
-    display block
-
-  .logo
-    width: 48px
-    height: 49px
-    background: url("../../images/logo@2x.png") no-repeat;
-    background-size: @width @height;
-
-  .links
-    display flex
-    iphoneDisabled()
-
-    a
-      color $black
-      font-size $small
-      line-height 20px
-      text-decoration none
-      margin-right 32px
-      cursor pointer
-
-      &:last-child
-        margin-right 0
