--- src/pages/Landing2.0/sections/Services/components/Service/index.styl
@@ -1,49 +1,0 @@
-@import '~styles/config.styl'
-
-.service
-  display flex
-  margin-bottom 200px
-
-  +iphone()
-    margin-bottom: 64px
-    display: block
-    width: 100%
-    padding: 0 40px
-    padding-bottom: 210px
-    position: relative
-
-  &:last-child
-    margin-bottom 160px
-
-    +iphone()
-      margin-bottom: 64px
-
-  .service-icons
-    display grid
-    grid-template-columns 1fr 1fr
-    grid-gap 70px
-
-    +iphone()
-      grid-gap 40px
-      position: absolute
-      bottom: 0
-      left: 0
-      width: 100%
-      padding: 0 40px
-
-    .service-icon
-      align-items center
-      display flex
-      flex-direction column
-
-      .icon-title
-        margin-top 20px
-        text-transform uppercase
-        color $gray
-        font-size $small
-        font-family "Roboto Mono", sans-serif
-        letter-spacing 2px
-        white-space nowrap
-
-        +iphone()
-          font-size: 12px
