--- src/pages/Landing2.0/sections/Services/index.styl
@@ -1,17 +1,0 @@
-@import '~styles/config.styl'
-
-.services
-  display flex
-  flex-direction column
-  align-items center
-  background-color $dark-blue
-
-  .section-heading
-    padding 120px 0 150px
-
-    +iphone()
-      padding: 64px 0
-      font-size: 28px
-      font-weight: $regular-weight
-      line-height: 1
-      letter-spacing 6px
